+++ dexterous_bioprosthesis_2021_raw_datasets/set_creators/np_signal_extractors/np_signal_extractor_ar.py
@@ -40,7 +40,7 @@ class NpSignalExtractorAr(NPSignalExtractor):
             m_params = [val for val in model.params]
             attribs += m_params
 
-        attribs = np.asanyarray(attribs)
+        attribs = np.asanyarray(attribs).astype(X.dtype)
         return attribs
 
     def attribs_per_column(self):
