+++ tests/raw_signals_creators/raw_signals_creator_sines_test.py
@@ -1,6 +1,8 @@
 
 import unittest
 
+import numpy as np
+
 from tests.raw_signals_creators.raw_signals_creator_test import RawSignalsCreatorTest
 from dexterous_bioprosthesis_2021_raw_datasets.raw_signals_creators.raw_signals_creator_sines import RawSignalsCreatorSines
 
@@ -8,8 +10,8 @@ class RawSignalsCreatorSinesTest(RawSignalsCreatorTest):
 
     __test__ = True
 
-    def get_creator(self):
-        return RawSignalsCreatorSines()
+    def get_creator(self,dtype=np.float32):
+        return RawSignalsCreatorSines(dtype=dtype)
 
 if __name__ == '__main__':
     unittest.main()
