+++ tests/raw_signals_spoilers/raw_signals_spoiler_interface_test.py
@@ -19,11 +19,11 @@ class RawSignalsSpoilerInterfaceTest(unittest.TestCase):
         """
         raise  unittest.SkipTest("Skipping")
     
-    def generate_sample_data(self,signal_number=10, column_number=3, samples_number=12)->RawSignals:
+    def generate_sample_data(self,signal_number=10, column_number=3, samples_number=12, dtype=np.double)->RawSignals:
         signals = RawSignals()
 
         for i in range(1,signal_number+1):
-            signals.append( RawSignal(signal=np.random.random( (samples_number,column_number))) )
+            signals.append( RawSignal(signal=np.random.random( (samples_number,column_number)).astype(dtype)) )
 
         return signals
